Code was strongly refined, VE version was separated

+++ bot/handlers/classes.py
@@ -3,13 +3,15 @@ from bot import students
 from bot import metrics
 from bot import top_notification
 
-from bot.constants import WEEKDAYS
-
 from bot.keyboards.classes import schedule_type
 from bot.keyboards.classes import certain_date_chooser
 
+from bot.helpers.constants import WEEKDAYS
+from bot.helpers.constants import LOADING_REPLIES
+
 from datetime import datetime
 from re import fullmatch
+from random import choice
 
 
 @kbot.message_handler(
@@ -26,7 +28,7 @@ def classes(message):
         if students[message.chat.id].another_group_number_schedule is None:
             kbot.send_message(
                 chat_id=message.chat.id,
-                text="Сайт kai.ru не отвечает🤷🏼‍♀️",
+                text="kai.ru не отвечает🤷🏼‍♀️",
                 disable_web_page_preview=True
             )
             
@@ -50,11 +52,19 @@ def one_day_schedule(callback):
     kbot.edit_message_text(
         chat_id=callback.message.chat.id,
         message_id=callback.message.message_id,
-        text=students[callback.message.chat.id].get_schedule(
-            type="classes",
-            weekday=int(callback.data[11:]),
-            next="next" in callback.data
-        ),
+        text=choice(LOADING_REPLIES),
+        disable_web_page_preview=True
+    )
+    
+    schedule = students[callback.message.chat.id].get_schedule(
+        type="classes",
+        next="next" in callback.data
+    )
+    
+    kbot.edit_message_text(
+        chat_id=callback.message.chat.id,
+        message_id=callback.message.message_id,
+        text=schedule[int(callback.data[11:]) - 1] if len(schedule) != 1 else schedule[0],
         parse_mode="Markdown",
         disable_web_page_preview=True
     )
@@ -84,19 +94,34 @@ def certain_date_schedule(callback):
 )
 @top_notification
 def weekly_schedule(callback):
-    kbot.delete_message(chat_id=callback.message.chat.id, message_id=callback.message.message_id)
+    kbot.edit_message_text(
+        chat_id=callback.message.chat.id,
+        message_id=callback.message.message_id,
+        text=choice(LOADING_REPLIES),
+        disable_web_page_preview=True
+    )
     
-    for weekday in WEEKDAYS:
-        kbot.send_message(
+    schedule = students[callback.message.chat.id].get_schedule(
+        type="classes",
+        next="next" in callback.data
+    )
+    
+    if len(schedule) == 1:
+        kbot.edit_message_text(
             chat_id=callback.message.chat.id,
-            text=students[callback.message.chat.id].get_schedule(
-                type="classes",
-                weekday=weekday,
-                next="next" in callback.data
-            ),
-            parse_mode="Markdown",
-            disable_web_page_preview=True
+            message_id=callback.message.message_id,
+            text=schedule[0]
         )
+    else:
+        kbot.delete_message(chat_id=callback.message.chat.id, message_id=callback.message.message_id)
+        
+        for weekday in WEEKDAYS:
+            kbot.send_message(
+                chat_id=callback.message.chat.id,
+                text=schedule[weekday - 1],
+                parse_mode="Markdown",
+                disable_web_page_preview=True
+            )
     
     students[callback.message.chat.id].another_group_number_schedule = None
     
@@ -105,33 +130,3 @@ def weekly_schedule(callback):
 
 @kbot.message_handler(func=lambda message: students[message.chat.id].previous_message == "/classes")
 def gs_classes(message): kbot.delete_message(chat_id=message.chat.id, message_id=message.message_id)
-
-
-@kbot.message_handler(
-    commands=["exams"],
-    func=lambda message: students[message.chat.id].previous_message is None
-)
-@metrics.increment("exams")
-def exams(message):
-    if fullmatch("[1-59][1-6][0-9][0-9]", message.text.replace("/exams ", "")):
-        students[message.chat.id].another_group_number_schedule = message.text.replace("/exams ", "")
-        
-        if students[message.chat.id].another_group_number_schedule is None:
-            kbot.send_message(
-                chat_id=message.chat.id,
-                text="Сайт kai.ru не отвечает🤷🏼‍♀️",
-                disable_web_page_preview=True
-            )
-            
-            students[message.chat.id].previous_message = None  # Gate System (GS)
-            
-            return
-
-    kbot.send_message(
-        chat_id=message.chat.id,
-        text=students[message.chat.id].get_schedule(type="exams"),
-        parse_mode="Markdown",
-        disable_web_page_preview=True
-    )
-    
-    students[message.chat.id].another_group_number_schedule = None
